+++ 05.Polymorphism/WildFarm/Models/Foods/Fruit.cs
@@ -7,7 +7,7 @@ namespace WildFarm.Foods
 {
     public class Fruit : Food
     {
-        public Fruit(int quantity) : base(quantity)
+        public Fruit(long quantity) : base(quantity)
         {
         }
     }
