+++ 05.Polymorphism/WildFarm/Models/Foods/Meat.cs
@@ -7,7 +7,7 @@ namespace WildFarm.Foods
 {
     public class Meat : Food
     {
-        public Meat(int quantity) : base(quantity)
+        public Meat(long quantity) : base(quantity)
         {
         }
     }
