+++ 05.Polymorphism/WildFarm/Models/Foods/Vegetable.cs
@@ -7,7 +7,7 @@ namespace WildFarm.Foods
 {
     public class Vegetable : Food
     {
-        public Vegetable(int quantity) : base(quantity)
+        public Vegetable(long quantity) : base(quantity)
         {
         }
     }
